Updated project structure | added screens logic | added play button to start game screen
